Large rewrite: Renamed namespace DWL.HTTP.Server.* to DWL.Server.*, This is a breaking change, clients need to be updated regarding this rename. Also completely rewrote DWLServer to be based on native TCPServer and no longer on Indy. Next commit will be a cleanup of the Indy source

+++ src/DWL.HTTP.Utils.pas
@@ -4,7 +4,7 @@ interface
 
 type
   TdwlHTTPUtils = record
-    class function StringTodwlhttpCommand(const HttpCommand: string): byte; static;
+    class function StringTodwlhttpCommand(const HttpCommand: string): integer; static;
     class function StatusCodeDescription(StatusCode: integer): string; static;
   end;
 
@@ -62,7 +62,7 @@ begin
   end;
 end;
 
-class function TdwlHTTPUtils.StringTodwlhttpCommand(const HttpCommand: string): byte;
+class function TdwlHTTPUtils.StringTodwlhttpCommand(const HttpCommand: string): integer;
 begin
   Result := IndexStr(HttpCommand, dwlhttpCommandToString);
 end;
